requirements.txt
numpy==2.1.1
PySDL2==0.9.16
pysdl2-dll==2.30.2
